feat: support `prove`
